Extract LegacyWorkspace fixture and add migration content-fidelity oracle.
Move the LegacyWorkspace fixture (and its helpers) out of
git-migration.test.ts into a shared __tests__/legacy-workspace.ts,
parameterized over storage so it can seed either mock storage or a real
Durable Object's storage. Add readDocFiles/expectHeadsMatchDoc: an oracle
that checks every migrated gadget head against an independent replay of
the recorded Yjs update log, applied in three representative unit tests.
